Cambios menores a energia_0

+++ src/metropolis.h
@@ -4,5 +4,5 @@ int metropolis(int *lattice, int n, float T, float *EM);
 int pick_site(int *lattice, int n);
 int suma_vecinos(int* lattice, int idx);
 int flip(int *lattice, int n, float T, int idx, float B float *EM);
-int energia_0(int *lattice,int n, int B);
+float energia_0(int *lattice,int n, float B);
 #endif
